Fix a few small issues with query logging.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@16070 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/gui/QvisHohlraumFluxQueryWidget.C
@@ -293,7 +293,7 @@ QvisHohlraumFluxQueryWidget::GetQueryParameters(MapNode &params)
     if (noerrors)
     {
         params["num_lines"] = nl;
-        params["divide_emis_by_absorb"] = divideFlag->isChecked();
+        params["divide_emis_by_absorb"] = (int)divideFlag->isChecked();
         params["ray_center"] = rc;
         params["radius"] = rad;
         params["theta"] = t;
